[♻️refactor] : 코드, 스타일링 폴더 구조 수정
